fix(rewise): Force useChat to reinitialize when chatId is changed
